+++ warehouse-service/warehouse-domain/warehouse-application-service/src/test/java/com/shopper/autos/system/warehouse/service/domain/WarehouseTestConstant.java
@@ -24,6 +24,6 @@ public class WarehouseTestConstant {
     public static final Integer PAGE = 0;
     public static final Integer SIZE = 1;
     public static final List<String> FIELDS = Arrays.asList("country", "state", "city");
-    public static final Integer TOTAL_RESULT = 1;
+    public static final Long TOTAL_RESULT = 1L;
 
 }
